added check for nvidia-container-toolkit/docker repo existence

+++ nv_driver_cuda.sh
@@ -52,20 +52,24 @@ fi
 
 echo -e "${YELLOW}Verifying nvidia-container-toolkit repository. Please wait...${ENDCOLOR}"
 sleep 1
-curl -fsSL https://nvidia.github.io/libnvidia-container/gpgkey | gpg --dearmor -o /usr/share/keyrings/nvidia-container-toolkit-keyring.gpg \
-&& curl -s -L https://nvidia.github.io/libnvidia-container/stable/deb/nvidia-container-toolkit.list | \
-sed 's#deb https://#deb [signed-by=/usr/share/keyrings/nvidia-container-toolkit-keyring.gpg] https://#g' | \
-tee /etc/apt/sources.list.d/nvidia-container-toolkit.list
+if [ -s /etc/apt/sources.list.d/nvidia-container-toolkit.list ]; then
+    curl -fsSL https://nvidia.github.io/libnvidia-container/gpgkey | gpg --dearmor -o /usr/share/keyrings/nvidia-container-toolkit-keyring.gpg
+    curl -s -L https://nvidia.github.io/libnvidia-container/stable/deb/nvidia-container-toolkit.list | \
+    sed 's#deb https://#deb [signed-by=/usr/share/keyrings/nvidia-container-toolkit-keyring.gpg] https://#g' | \
+    tee /etc/apt/sources.list.d/nvidia-container-toolkit.list
+fi
 
 echo -e "${YELLOW}Verifying Docker repository. Please wait...${ENDCOLOR}"
 sleep 1
-mkdir -p /etc/apt/keyrings && chmod 0755 /etc/apt/keyrings
-curl -kfsSL https://download.docker.com/linux/ubuntu/gpg | gpg --dearmor -o /etc/apt/keyrings/docker.gpg
-chmod a+r /etc/apt/keyrings/docker.gpg
-echo \
- "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
- $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
- tee /etc/apt/sources.list.d/docker.list > /dev/null
+if [ ! -s /etc/apt/keyrings/docker.gpg ] || [ ! -s /etc/apt/sources.list.d/docker.list ]; then
+    mkdir -p /etc/apt/keyrings && chmod 0755 /etc/apt/keyrings
+    curl -kfsSL https://download.docker.com/linux/ubuntu/gpg | gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+    chmod a+r /etc/apt/keyrings/docker.gpg
+    echo \
+    "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
+    $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
+    tee /etc/apt/sources.list.d/docker.list > /dev/null
+fi
 
 echo -e "${YELLOW}Installing CUDA toolkit, NVIDIA driver, docker, and nvidia-container-toolkit. Please wait...${ENDCOLOR}"
 apt update
